Nostr/requirements.txt
requests==2.25.1
numpy==1.25
python-dotenv==1.0.0
ecdsa==0.14.1
cffi>=1.15.0
cryptography>=37.0.4
secp256k1>=0.14.0
websocket-client>=1.3.3
pytest>=7.2.0
openai
psycopg2_binary==2.9.6

